backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.26.0
pandas>=2.0.3
scipy>=1.11.3
scikit-learn>=1.3.0
torch>=2.0.1
torchvision>=0.15.2
timm>=0.9.2
python-multipart>=0.0.6
pillow>=10.0.0
tqdm>=4.66.1 